+++ NEWS.rst
@@ -109,6 +109,9 @@ unambiguously guessed from the atom name, in accordance with PDB structures.
 Bio.PDB entities now have a ``center_of_mass()`` method that calculates either
 centers of gravity or geometry.
 
+Addition of `Bio.PDB.read()` function that supports reading pdb, pdb.gz, cif,
+cif.gz, mmtf.
+
 New method ``disordered_remove()`` implemented in Bio.PDB DisorderedAtom and
 DisorderedResidue to remove children.
 
@@ -144,6 +147,7 @@ possible, especially the following contributors:
 - Suyash Gupta
 - Vini Salazar (first contribution)
 - Leighton Pritchard
+- Simon Duerr (first contribution)
 
 4 September 2020: Biopython 1.78
 ================================